Play game against UCI
Thanks to Phoenix's Socket module, we are now able
to connect to a WS channel and play against Stockfish.

For now, users only play on the 'white' side and there is no winner.
This will be fixed in the next commits as the priority was to play
against a UCI.